native: fix LP64 return-type mismatches (getResSize long->s32, render-height u16)
- JKRFileLoader::getResSize: base virtual returned 'long' (64-bit on LP64) while
  the JKRArchive override returns s32; on GC long==s32 so they matched. Make the
  base s32 to match the override (16 files were 'conflicting return type').
- SMSGetGameRender{Height,Width}: cameralib.hpp declared s16, MapDraw.cpp int,
  but the canonical defn (Resolution) is u16. A host TU seeing two return types
  errors 'ambiguating new declaration'. Unify on u16 (values 448/640 fit all;
  the int/s16 variants were GC fakematch quirks).

+++ src/Map/MapDraw.cpp
@@ -12,8 +12,12 @@
 // NOTE: Yes, this is an ODR violation, and I'm pretty sure they had it
 // in the original code as well, because everywhere else these functions
 // return a u16, but here they MUST return an int for some reason???
-int SMSGetGameRenderHeight();
-int SMSGetGameRenderWidth();
+// Native build: must match the canonical u16 decl (System/Resolution.hpp /
+// Camera/cameralib.hpp) -- a TU seeing both an int and a u16 decl errors with
+// "ambiguating new declaration". The original's `int` here was a GC codegen
+// fakematch quirk with no bearing on the value (448/640 fit either type).
+u16 SMSGetGameRenderHeight();
+u16 SMSGetGameRenderWidth();
 
 void TMapDrawWall::perform(u32 param_1, JDrama::TGraphics* param_2)
 {
